fix: sections table modified field

+++ database/migrations/2022_11_24_230747_create_sections_table.php
@@ -16,8 +16,8 @@ return new class extends Migration
         Schema::create('sections', function (Blueprint $table) {
             $table->id();
             $table->string('name', 100)->unique();
-            $table->unsignedInteger('my_class_id')->unique();
-            $table->unsignedInteger('teacher_id')->nullable();
+            $table->unsignedBigInteger('my_class_id')->unique();
+            $table->unsignedBigInteger('teacher_id')->nullable();
             $table->tinyInteger('active')->default(0);
             $table->timestamps();
         });
